+++ firmware/core/src/services/storage_service.cpp
@@ -12,7 +12,7 @@ std::string StorageService::nsKey(const std::string& bundleId) {
     uint32_t h = 5381;
     for (char c : bundleId) h = ((h << 5) + h) + (uint8_t)c;
     char buf[9];
-    snprintf(buf, sizeof(buf), "%08x", h);
+    snprintf(buf, sizeof(buf), "%08x", (unsigned int)h);
     return buf;
 }
